Add files via upload
This is literally just my entire file. Changes are within every block2 file and ex3 file. Displayed login and registration forms when you click on menu in block2. Added drag and drop function in ex3. --> built ex3 on block2 code, but it needs to be the same as in exercise one, so merge new code into the old folder.